fdd_pipeline/requirements.txt
pydantic>=2.0.0
requests>=2.25.0
boto3>=1.24.0
pandas>=2.0.0
plotly>=5.15.0
dash>=2.14.0
python-dotenv>=1.0.0
huggingface-hub>=0.17.0
PyMuPDF>=1.22.0
instructor
rapidfuzz
markdownify
mistral
generativeai
google-generativeai
sentence-transformers>=2.2.2
numpy>=1.24.3
scipy>=1.10.0
rapidfuzz>=3.0.0
docling-core
docling
mlx-vlm
rapidfuzz


# Development tools
pytest>=7.4.3
black>=23.12.1
flake8>=6.1.0
mypy>=1.7.1
ruff>=0.1.13
pytesseract
pillow
opencv-python
requests>=2.28.0
supabase>=0.0.2
typing-extensions>=4.2.0 

# LLM dependencies
openai>=1.1.0
anthropic>=0.39.0
scikit-learn>=1.2.2
boto3
pdfplumber
nltk
spacy
pandas
PyPDF2
torch
transformers
huggingface-hub